==> 3_mean_SE.Rmd <==
---
title: "Mean / standard error"
author: "Francisca Rodriguez"
date: "11/4/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**Pre-decontamination**
```{r}
library(phyloseq)
library(dplyr)

#load data
ps <- readRDS("./intermediate_data/ps.rds")

#create a richness function
richness <- function(v) {
  y <- sum(v > 0)
  return(y)
}

#create a standard error function 
se <- function(s) {
  x <- sd(s)/sqrt(length(s))
  return(x)
}

```

**Subset data**
```{r}
#initialize variables + lists for loop
sd <- as.data.frame(sample_data(ps))
categories <- unique(sd$SampleType)
all_samples <- list()

for (a_cat in categories) {
  
  subset_ps <- subset_samples(ps, sample_data(ps)$SampleType==a_cat)
  pruned_ps <- prune_taxa(taxa_sums(subset_ps) > 0, subset_ps)

  all_samples[[a_cat]] <- subset_ps  #phyloseq object list of data by sample type
  
}

```


**Calculate richness / mean / se / replicate number (n)**
```{r}
library(dplyr)

data <- c("slurry", "post_inoc_post_stress", "post_inoc_pre_stress", "pre_inoc", "di_water", "DNA_neg", "i6" )
sampling <- c("Type", "Water", "SampleType", "SampleType", "SampleType", "SampleType", "SampleType" )

total_data <- data.frame()

#create table with all sample type data
for (i in 1:length(data)){
  x_asv <- subset_samples(all_samples[[data[i]]])
  meta <- data.frame(sample_data(x_asv))[,sampling[i]]
  x_asv <- as.data.frame(otu_table(x_asv))
  meta_x <- data.frame(meta = as.factor(meta))
  
  richness_value <- apply(x_asv, 1, richness)
  
  meta_x$richness <- richness_value
  meta_x <- meta_x %>% group_by(meta) %>% mutate(rep = n())
  meta_x$data <- data[i]
  meta_x$sampling <- sampling[i]
  
  if (i == 1 | i == 2 | i == 5){
    if (i == 1 | i == 5){meta_x$stats <- "split"}
    if (i == 2){meta_x$stats <- "apply"}
  }
  else {meta_x$stats <- "basic"}
  
  total_data <- rbind(total_data, meta_x)
}

sorting_types <- unique(total_data$stats)
summary_data <- data.frame()

for (i in sorting_types){
  
  #calculate mean/se/rep. for slurry and di water (only one rep. each)
  if (i == "split"){
    x <- subset(total_data, stats == "split")
    sample_type <- x$meta 
    richness <- x$richness
    dat <- data.frame(sample_type)
    dat$replicate <- 1
    dat$se <- 0
    dat$richness <- richness
    summary_data <- rbind(summary_data, dat)
  }
  
  #calculate mean/se/rep. for samples that don't require subsetting by category
  if (i == "basic"){
    x <- subset(total_data, stats == "basic")
    dat <- x %>% group_by(data) %>% summarize(
                                           replicate = n(),
                                           se = se(richness),
                                           richness = mean(richness))
    dat <- rename(dat, c("sample_type" = "data"))
    summary_data <- rbind(summary_data, dat)
  }

  #subset post_inoc_post_stress by incoulation and water
   if (i == "apply"){
    x <- subset(total_data, stats == "apply")
    x$water <- sample_data(subset_samples(all_samples[["post_inoc_post_stress"]]))$Water
    x$inoculation <- sample_data(subset_samples(all_samples[["post_inoc_post_stress"]]))$Inoculation
    x$sample_type <- as.factor(paste("W", x$water, "I", x$inoculation, sep=""))
    
    dat <- x %>% group_by(sample_type) %>% summarize(
                                           replicate = n(),
                                           se = se(richness),
                                           richness = mean(richness))
    summary_data <- rbind(summary_data, dat)
  }
  
}

#export as .csv
write.csv(summary_data, "./output/meanSE_total.csv")

```
