==> 6_alpha_beta_curves.Rmd <==
---
title: "Alpha / beta diversity curves"
author: "Francisca Rodriguez"
date: "11/4/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**Post-decontamination**
```{r}
#load libraries / data
library(phyloseq)
library(ggplot2)
library(decontam)
library(vegan)

ps.noncontam <- readRDS("./intermediate_data/ps.noncontam.rds")

```

**Subset by sample type**
```{r}
#initialize variables + lists for loop
sd <- as.data.frame(sample_data(ps.noncontam))
categories <- unique(sd$SampleType)
all_samples_pruned_asv <- list()
all_samples_pruned <- list()

for (a_cat in categories) {
  
  subset_ps <- subset_samples(ps.noncontam, sample_data(ps.noncontam)$SampleType==a_cat)
  pruned_ps <- prune_taxa(taxa_sums(subset_ps) > 0, subset_ps)
  asv <- as.data.frame(otu_table(pruned_ps))
  sample_names(pruned_ps) <- paste(a_cat, sprintf('%0.4d', 1:length(sample_names(pruned_ps))), sep = "") # generate the row names 

  all_samples_pruned_asv[[a_cat]] <- asv    #asv tables of data by sample type
  all_samples_pruned[[a_cat]] <- subset_ps  #phyloseq object list of data by sample type
  
}

```

**Subset post-inoculation pre-stress**
```{r}
#establish inoculation names for loops
inoc_names <- c("Inoc1", "Inoc2", "Inoc3", "Inoc4", "Inoc5", "Inoc6")
inocs_vec <- c("1", "2", "3", "4", "5", "6")
prestress_names <- sapply(inoc_names, function(x) paste(x))
prestress_asvs <- list()
prestress <- list()

#loop to subset
for (inoc in inocs_vec) {
  
  subset_ps <- subset_samples(all_samples_pruned[["post_inoc_pre_stress"]], 
                              Inoculation == inoc)
  pruned_ps <- prune_taxa(taxa_sums(subset_ps) > 0, subset_ps)
  asv <- as.data.frame(otu_table(pruned_ps))
  
  prestress_asvs[[inoc]] <- asv   #asv list of post-inoc pre-stress data
  prestress[[inoc]] <- pruned_ps  #phyloseq object list of pre-stress data
}

#add names 
names(prestress) <- prestress_names
names(prestress_asvs) <- prestress_names

#subset as needed for analysis
post_inoc_pre_stress_i1 <- as.data.frame(otu_table(prestress[["Inoc1"]]))
post_inoc_pre_stress_i2 <- as.data.frame(otu_table(prestress[["Inoc2"]]))
post_inoc_pre_stress_i3 <- as.data.frame(otu_table(prestress[["Inoc3"]]))
post_inoc_pre_stress_i4 <- as.data.frame(otu_table(prestress[["Inoc4"]]))
post_inoc_pre_stress_i5 <- as.data.frame(otu_table(prestress[["Inoc5"]]))
post_inoc_pre_stress_i6 <- as.data.frame(otu_table(prestress[["Inoc6"]]))

```

**Subset post-inoculation post-stress**
```{r}
##establish inoculation names for loops
inoc_names <- c("Inoc1", "Inoc2", "Inoc3", "Inoc4", "Inoc5")
inocs_vec <- c("1", "2", "3", "4", "5")
  
## Water 1 ## 
water1 <- list()
water1_names <- sapply(inoc_names, function(x) paste(x, "Water1"))

for (inoc in inocs_vec) {
  a_ps <- subset_samples(all_samples_pruned[["post_inoc_post_stress"]], 
                         Inoculation == inoc & Water == "1")
  a_pr <- prune_taxa(taxa_sums(a_ps) > 0, a_ps)
  water1[[inoc]] <- a_pr
}

names(water1) <- water1_names

## Water 2 ## 
water2 <- list()
water2_names <- sapply(inoc_names, function(x) paste(x, "Water2"))

for (inoc in inocs_vec) {
  a_ps <- subset_samples(all_samples_pruned[["post_inoc_post_stress"]], 
                         Inoculation == inoc & Water == "2")
  a_pr <- prune_taxa(taxa_sums(a_ps) > 0, a_ps)
  water2[[inoc]] <- a_pr
}

names(water2) <- water2_names

## Water 3 ## 
water3 <- list()
water3_names <- sapply(inoc_names, function(x) paste(x, "Water3"))

for (inoc in inocs_vec) {
  a_ps <- subset_samples(all_samples_pruned[["post_inoc_post_stress"]], 
                         Inoculation == inoc & Water == "2")
  a_pr <- prune_taxa(taxa_sums(a_ps) > 0, a_ps)
  water3[[inoc]] <- a_pr
}

names(water3) <- water3_names

# make big list for all post inoc post stress treatments
post_inoc_post_stress <- c(water1, water2, water3)

#inoculation 1
post_inoc_post_stress_i1_w1 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc1 Water1"]]))
post_inoc_post_stress_i1_w2 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc1 Water2"]]))
post_inoc_post_stress_i1_w3 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc1 Water3"]]))

#inoculation 2
post_inoc_post_stress_i2_w1 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc2 Water1"]]))
post_inoc_post_stress_i2_w2 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc2 Water2"]]))
post_inoc_post_stress_i2_w3 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc2 Water2"]]))

#inoculation 3
post_inoc_post_stress_i3_w1 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc3 Water1"]]))
post_inoc_post_stress_i3_w2 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc3 Water2"]]))
post_inoc_post_stress_i3_w3 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc3 Water3"]]))

#inoculation 4
post_inoc_post_stress_i4_w1 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc4 Water1"]]))
post_inoc_post_stress_i4_w2 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc4 Water2"]]))
post_inoc_post_stress_i4_w3 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc4 Water3"]]))

#inoculation 5
post_inoc_post_stress_i5_w1 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc5 Water1"]]))
post_inoc_post_stress_i5_w2 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc5 Water2"]]))
post_inoc_post_stress_i5_w3 <- as.data.frame(otu_table(post_inoc_post_stress[["Inoc5 Water3"]]))
```

**Species accumulation curves - beta diversity**
```{r}
par(mfrow=c(4,5))

##POST INOC PRE STRESS #####################################################
accurve_prei1 <- specaccum(post_inoc_pre_stress_i1, method = "random", permutations = 400)
plot(accurve_prei1$sites, accurve_prei1$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc pre-stress ASVs - Inoculation 1")

accurve_prei2 <- specaccum(post_inoc_pre_stress_i2, method = "random", permutations = 400)
plot(accurve_prei2$sites, accurve_prei2$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc pre-stress ASVs - Inoculation 2")

accurve_prei3 <- specaccum(post_inoc_pre_stress_i3, method = "random", permutations = 400)
plot(accurve_prei3$sites, accurve_prei3$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc pre-stress ASVs - Inoculation 3")

accurve_prei4 <- specaccum(post_inoc_pre_stress_i4, method = "random", permutations = 400)
plot(accurve_prei4$sites, accurve_prei4$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc pre-stress ASVs - Inoculation 4")

accurve_prei5 <- specaccum(post_inoc_pre_stress_i5, method = "random", permutations = 400)
plot(accurve_prei5$sites, accurve_prei5$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc pre-stress ASVs - Inoculation 5")

accurve_prei6 <- specaccum(post_inoc_pre_stress_i6, method = "random", permutations = 400)
plot(accurve_prei6$sites, accurve_prei6$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc pre-stress ASVs - Inoculation 6")


##POST INOC POST STRESS #####################################################

accurve_posti1w1 <- specaccum(post_inoc_post_stress_i1_w1, method = "random", permutations = 400)
plot(accurve_posti1w1$sites, accurve_posti1w1$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i1 x w1")

accurve_posti1w2 <- specaccum(post_inoc_post_stress_i1_w2, method = "random", permutations = 400)
plot(accurve_posti1w2$sites, accurve_posti1w2$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i1 x w2")

accurve_posti1w3 <- specaccum(post_inoc_post_stress_i1_w3, method = "random", permutations = 400)
plot(accurve_posti1w3$sites, accurve_posti1w3$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i1 x w3")

accurve_posti2w1 <- specaccum(post_inoc_post_stress_i2_w1, method = "random", permutations = 400)
plot(accurve_posti2w1$sites, accurve_posti2w1$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i2 x w1")

accurve_posti2w2 <- specaccum(post_inoc_post_stress_i2_w2, method = "random", permutations = 400)
plot(accurve_posti2w2$sites, accurve_posti2w2$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i2 x w2")

accurve_posti2w3 <- specaccum(post_inoc_post_stress_i2_w3, method = "random", permutations = 400)
plot(accurve_posti2w3$sites, accurve_posti2w3$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i2 x w3")

accurve_posti3w1 <- specaccum(post_inoc_post_stress_i3_w1, method = "random", permutations = 400)
plot(accurve_posti3w1$sites, accurve_posti3w1$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i3 x w1")

accurve_posti3w2 <- specaccum(post_inoc_post_stress_i3_w2, method = "random", permutations = 400)
plot(accurve_posti3w2$sites, accurve_posti3w2$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i3 x w2")

accurve_posti3w3 <- specaccum(post_inoc_post_stress_i3_w3, method = "random", permutations = 400)
plot(accurve_posti3w3$sites, accurve_posti3w3$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i3 x w3")

accurve_posti4w1 <- specaccum(post_inoc_post_stress_i4_w1, method = "random", permutations = 400)
plot(accurve_posti4w1$sites, accurve_posti4w1$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i4 x w1")

accurve_posti4w2 <- specaccum(post_inoc_post_stress_i4_w2, method = "random", permutations = 400)
plot(accurve_posti4w2$sites, accurve_posti4w2$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress - i4 x w2")

accurve_posti4w3 <- specaccum(post_inoc_post_stress_i4_w3, method = "random", permutations = 400)
plot(accurve_posti4w3$sites, accurve_posti4w3$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress - i4 x w3")

accurve_posti5w1 <- specaccum(post_inoc_post_stress_i5_w1, method = "random", permutations = 400)
plot(accurve_posti5w1$sites, accurve_posti5w1$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i5 x w1")

accurve_posti5w2 <- specaccum(post_inoc_post_stress_i5_w2, method = "random", permutations = 400)
plot(accurve_posti5w2$sites, accurve_posti5w2$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i5 x w2")

accurve_posti5w3 <- specaccum(post_inoc_post_stress_i5_w3, method = "random", permutations = 400)
plot(accurve_posti5w3$sites, accurve_posti5w3$richness,
     xlab = "Number of sample",
     ylab = "Species richness",
     main = "Post-inoc post-stress ASVs - i5 x w3")

```

**Sequencing depth curves - alpha diversity**
```{r}
par(mfrow=c(4,5))

##POST INOC PRE STRESS#######################################################

#post_inoc_pre_stress_i1
rarecurve(post_inoc_pre_stress_i1,
          step = 1, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc pre-stress ASVs - Inoculation 1")

#post_inoc_pre_stress_i2
rarecurve(post_inoc_pre_stress_i2, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc pre-stress ASVs - Inoculation 2")

#post_inoc_pre_stress_i3
rarecurve(post_inoc_pre_stress_i3, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size",
          ylab = "ASV", 
          main = "Post-inoc pre-stress ASVs - Inoculation 3")

#post_inoc_pre_stress_i4
rarecurve(post_inoc_pre_stress_i4, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc pre-stress ASVs - Inoculation 4")

#post_inoc_pre_stress_i5
rarecurve(post_inoc_pre_stress_i5, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc pre-stress ASVs - Inoculation 5")

#post_inoc_pre_stress_i6
rarecurve(post_inoc_pre_stress_i6, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc pre-stress ASVs - Inoculation 6")

##POST INOC POST STRESS#######################################################

##INOCULATION 1
#post_inoc_post_stress_i1_w1
rarecurve(post_inoc_post_stress_i1_w1, 
          step = 20,
          label = FALSE,
          col = "blue", 
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i1 x w1")

#post_inoc_post_stress_i1_w2
rarecurve(post_inoc_post_stress_i1_w2, 
          step = 20, 
          col = "blue",
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i1 x w2")

#post_inoc_post_stress_i1_w3
rarecurve(post_inoc_post_stress_i1_w3, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i1 x w3")

##INOCULATION 2
#post_inoc_post_stress_i2_w1
rarecurve(post_inoc_post_stress_i2_w1, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i2 x w1")

#post_inoc_post_stress_i2_w2
rarecurve(post_inoc_post_stress_i2_w2, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i2 x w2")

#post_inoc_post_stress_i2_w3
rarecurve(post_inoc_post_stress_i2_w3, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i2 x w3")

##INOCULATION 3
#post_inoc_post_stress_i3_w1
rarecurve(post_inoc_post_stress_i3_w1, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i3 x w1")

#post_inoc_post_stress_i3_w2
rarecurve(post_inoc_post_stress_i3_w2, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i3 x w2")

#post_inoc_post_stress_i3_w3
rarecurve(post_inoc_post_stress_i3_w3, 
          step = 20, 
          col = "blue", 
          cex = 0.6, 
          label = FALSE,
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i3 x w3")

##INOCULATION 4
#post_inoc_post_stress_i4_w1
rarecurve(post_inoc_post_stress_i4_w1, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i4 x w1")

#post_inoc_post_stress_i4_w2
rarecurve(post_inoc_post_stress_i4_w2, 
          step = 20, 
          col = "blue", 
          cex = 0.6, 
          label = FALSE,
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i4 x w2")

#post_inoc_post_stress_i4_w3
rarecurve(post_inoc_post_stress_i4_w3, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i4 x w3")


##INOCULATION 5 
#post_inoc_post_stress_i5_w1
rarecurve(post_inoc_post_stress_i5_w1, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i5 x w1")

#post_inoc_post_stress_i5_w2
rarecurve(post_inoc_post_stress_i5_w2, 
          step = 20, 
          col = "blue", 
          cex = 0.6, 
          label = FALSE,
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i5 x w2")

#post_inoc_post_stress_i5_w3
rarecurve(post_inoc_post_stress_i5_w3, 
          step = 20, 
          col = "blue", 
          label = FALSE,
          cex = 0.6, 
          xlab = "Sample Size", 
          ylab = "ASV", 
          main = "Post-inoc post-stress ASVs - i5 x w3")

```
